+++ src/pid_control.cpp
@@ -11,57 +11,82 @@
 
 PidController pid;
 
+static uint8_t calc_digits(float n)
+{
+    if (n == 0) {
+        return 0;
+    }
+    uint8_t digits = 1;
+    for(;;) {
+        if (n >= 10) {
+            break;
+        }
+        n *= 10;
+        if (++digits >= 5) {
+            break;
+        }
+    }
+    return digits;
+}
+
+static uint32_t pow10_int(uint8_t n) {
+    uint32_t res = 10;
+    while(n--) {
+        res *= 10;
+    }
+    return res;
+}
+
 void PidController::updatePidValue(PidConfigEnum pid, int16_t steps)
 {
     auto &value = _settings[pid];
-    value += (KNOB_GET_VALUE(steps, KNOB_PID_SPEED) * 0.001);
+    auto multiplier = std::clamp(10.0 / pow10_int(value == 0 ? 5 : calc_digits(value)), 0.00001, 1.0);
+    value += (KNOB_GET_VALUE(steps, KNOB_PID_SPEED) * multiplier);
+    if (value < 0) {
+        value = 0;
+    }
 }
 
-uint8_t PidController::getDutyCycle() const
+void PidController::printValues(Print &output, bool displayOM) const
 {
-    return _dutyCycle >> kDutyCycleShift;
-}
-
-void PidController::printValues(Print &buffer) const
-{
-    char temp[80];
-    PrintBuffer buf(temp, sizeof(temp));
-
-    buf.printTrimmed(_settings.Kp);
-    buf.print(' ');
-    buf.printTrimmed(_settings.Ki);
-    buf.print(' ');
-    buf.printTrimmed(_settings.Kd);
-    buf.print(' ');
-    buf.printTrimmed(_settings.OutputMultiplier);
-
-    buffer.print(F("PID "));
-    buffer.println(buf.getBuffer());
+    output.print(F("PID "));
+    output.print(_settings.Kp, calc_digits(_settings.Kp));
+    output.print(' ');
+    output.print(_settings.Ki, calc_digits(_settings.Ki));
+    output.print(' ');
+    output.print(_settings.Kd, calc_digits(_settings.Kd));
+    if (displayOM) {
+        output.print(' ');
+        output.print(_settings.OutputMultiplier, calc_digits(_settings.OutputMultiplier));
+    }
+    output.println();
 }
 
 void PidController::displayPidSettingsMenu(PidConfigEnum highlight) const
 {
-    char temp[32];
-    PrintBuffer buf(temp, sizeof(temp));
+    if (highlight < PidConfigEnum::SAVE) {
+        display_print_hl((highlight == PidConfigEnum::KP), F("Kp"));
+        display.println(_settings.Kp, calc_digits(_settings.Kp));
+    }
 
-    display_print_hl((highlight == PidConfigEnum::KP), F("Kp"));
-    buf.clearPrintTrimmed(_settings.Kp);
-    display.println(buf.getBuffer());
-
-    display_print_hl((highlight == PidConfigEnum::KI), F("Ki"));
-    buf.clearPrintTrimmed(_settings.Ki);
-    display.println(buf.getBuffer());
+    if (highlight < PidConfigEnum::RESTORE) {
+        display_print_hl((highlight == PidConfigEnum::KI), F("Ki"));
+        display.println(_settings.Ki, calc_digits(_settings.Ki));
+    }
 
     display_print_hl((highlight == PidConfigEnum::KD), F("Kd"));
-    buf.clearPrintTrimmed(_settings.Kd);
-    display.println(buf.getBuffer());
+    display.println(_settings.Kd, calc_digits(_settings.Kd));
 
-    display_print_hl((highlight == PidConfigEnum::KD), F("OpMp"));
-    buf.clearPrintTrimmed(_settings.OutputMultiplier);
-    display.println(buf.getBuffer());
+    display_print_hl((highlight == PidConfigEnum::OM), F("Mul"));
+    display.println(_settings.OutputMultiplier, calc_digits(_settings.OutputMultiplier));
 
-    display_print_hl((highlight == PidConfigEnum::SAVE), F("Save"));
-    display_print_hl((highlight == PidConfigEnum::RESTORE), F("Restore"));
+    if (highlight >= PidConfigEnum::SAVE) {
+        display_print_hl((highlight == PidConfigEnum::SAVE), F("Save"), '\n');
+    }
+
+    if (highlight >= PidConfigEnum::RESTORE) {
+        display_print_hl(true, F("Restore"));
+    }
 }
 
 void PidController::reset()
@@ -73,34 +98,29 @@ void PidController::reset()
     _lastUpdate.start();
 }
 
-#if HAVE_GCC_OPTIMIZE_O3
-#    pragma GCC optimize("O3")
-#endif
+static inline float calc_delta_t(uint16_t time)
+{
+    return 1.0 / (10000000.0 / time);
+}
 
 void PidController::updateTicks(int32_t ticks)
 {
     auto _micros = micros();
-    float delta_t = 100000.0 / (float)_lastUpdate.getTime(_micros);
-    // 1.0 / (1000 * 1000.0 * _lastUpdate.getTime(_micros));
+    float delta_t; // gets calculated on demand
+    uint16_t time = _lastUpdate.getTime(_micros);
     _lastUpdate.start(_micros);
 
     float error = ticks - data.getSetPointRPMTicks(); // values are inverted
-    // error *= std::min<float>(1.0, delta_t);
-    // delta_t /= 1000;
-
-    // Serial.print(ticks);
-    // Serial.print(' ');
-    // Serial.print(data.getSetPointRPMTicks());
-    // Serial.print(' ');
-    // Serial.println(error);
-
-
     float output = _settings.Kp * error;
     if (_settings.Ki) {
+        delta_t = calc_delta_t(time);
         _integral = _integral + (error * delta_t);
         output += _settings.Ki * _integral;
     }
     if (_settings.Kd) {
+        if (!_settings.Ki) {
+            delta_t = calc_delta_t(time);
+        }
         float derivative = (error - _previousError) / delta_t;
         output += _settings.Kd * derivative;
         _previousError = error;
@@ -108,33 +128,11 @@ void PidController::updateTicks(int32_t ticks)
 
     output *= pid.getPidValues().OutputMultiplier;
 
-    // // // the output is divided by ((1 << kDutyCycleShift) * 32) = 4096 before added to the duty cycle 0-255
-    // int32_t tmp = output;
-    // tmp >>= 5;
-
-    uint8_t dutyCycleBefore = current_limit.getDutyCycle(getDutyCycle());
     _dutyCycle = std::clamp<int32_t>(_dutyCycle + output, MIN_DUTY_CYCLE_PID << kDutyCycleShift, MAX_DUTY_CYCLE_PID << kDutyCycleShift);
+
+    // apply current limit
     uint8_t dutyCycle8 = current_limit.getDutyCycle(getDutyCycle());
-
-    // Serial.print(dutyCycleBefore);
-    // Serial.print(' ');
-    // Serial.println(dutyCycle8);
-
-    #if HAVE_PID_CONTROLLER_STATS
-        if (dutyCycleBefore != dutyCycle8) {
-            setMotorPWM(dutyCycle8);
-            _stats.add(RPM_SENSE_TICKS_TO_RPM(ticks), ui_data.getRpm(), data.getSetPointRPM(), dutyCycle8);
-        }
-        else {
-            if (_stats._state == StatsState::RUNNING && ui_data.getRpm() == data.getSetPointRPM()) {
-                _stats.stop();
-                motor.stop(MotorStateEnum::OFF);
-            }
-        }
-    #else
-        if (dutyCycleBefore != dutyCycle8) {
-            // setMotorPWM();
-            setMotorPWM_timer(dutyCycle8);
-        }
-    #endif
+    if (getMotorPWM_timer() != dutyCycle8) {
+        setMotorPWM_timer(dutyCycle8);
+    }
 }